+++ NEWS.md
@@ -4,6 +4,7 @@
 
 #### Bug fixes
 - Fix str representation of multipoles
+- Fix implicit plotting of `Series` for `matplotlib >= 3.5.2`
 
 ## Version 1.3.4 (22 April 2022)